s3-cam/raspberry-pi-s3-cam/requirements.txt
Adafruit-DHT==1.3.2
certifi==2018.4.16
chardet==3.0.4
cryptography==1.7.1
dweepy==0.3.0
enum34==1.1.6
futures==3.2.0
idna==2.6
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
meld3==1.0.2
numpy==1.12.1
picamera==1.13
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
RPi.GPIO==0.6.3
SecretStorage==2.3.1
six==1.10.0
supervisor==3.3.1
tinys3==0.1.12
urllib3==1.22
